requirements.txt
matplotlib==3.8.2
pandas==2.2.1
numpy==1.26.3
streamlit==1.32.2
py-trello==0.19.0
requests==2.31.0
